Develope

+++ routes/api.php
@@ -13,28 +13,30 @@ Route::post('/register', [AuthController::class, 'register']);
 Route::post('/login', [AuthController::class, 'login']);
 // get all users
 Route::get('/users', [AuthController::class, 'index']);
-// get specified user
+// get specified user by id
 Route::get('/users/{id}', [AuthController::class, 'show']);
 // search user by name
 Route::get('/users/search/{name}', [AuthController::class, 'search']);
 
 // protected routes
 Route::group(['middleware' => ['auth:sanctum']], function () {
-    // create new url list
-    Route::post('/users/{id}', [AuthController::class, 'url_post']);
-    // change the informations of user by id
-    Route::put('/users/{id}', [AuthController::class, 'update']);
-    // change the informations of url table by id
-    Route::put('/users/{id}/urls/{url_id}', [AuthController::class, 'url_update']);
-    // delete item from url table
-    Route::delete('/users/{id}/urls/{url_id}', [AuthController::class, 'url_destroy']);
-    // delete item from users table
-    Route::delete('/users/{id}', [AuthController::class, 'destroy']);
-    // user logout
+    // create new url 
+    Route::post('/users/{id}', [AuthController::class, 'url_post'])->middleware('auth', 'check_user_ownership');
+    // change the information of user by id
+    Route::put('/users/{id}', [AuthController::class, 'update'])->middleware('auth', 'check_user_ownership');
+    // change the informations of url by id
+    Route::put('/users/{id}/urls/{url_id}', [AuthController::class, 'url_update'])->middleware('auth', 'check_user_ownership');
+    // delete selected user by id
+    Route::delete('/users/{id}', [AuthController::class, 'destroy'])->middleware('auth', 'check_user_ownership');
+    // delete selected url by id
+    Route::delete('/users/{id}/urls/{url_id}', [AuthController::class, 'url_destroy'])->middleware('auth', 'check_user_ownership');
+    // user logout and token destroy
     Route::post('/logout', [AuthController::class, 'logout']);
     
     // // sent a new record to accounts table
     // Route::post('/accounts', [AccountController::class, 'store']);
+
+
 });
 
 Route::middleware('auth:api')->get('/user', function (Request $request) {
